+++ webui/src/ElectronApp.tsx
@@ -5,41 +5,12 @@ import { SatelliteHeader } from './Header'
 import { SettingsForm } from './SettingsForm'
 import { ConnectionStatus } from './ConnectionStatus'
 import { MyErrorBoundary } from './Util/ErrorBoundary'
-import { useCallback, useEffect, useState } from 'react'
+import { useCallback } from 'react'
 import { usePoller } from './Util/usePoller'
-import type { ApiConfigData, ApiStatusResponse } from '../../src/apiTypes'
-import { SaveApiConfigData } from './Api/types'
+import type { ApiStatusResponse } from '../../src/apiTypes'
+import { useIpcConfigApi } from './Api/ipc'
 
 const STATUS_POLL_INTERVAL = 2000
-const CONFIG_POLL_INTERVAL = 5000
-
-function useIpcConfigApi(): {
-	currentConfig: ApiConfigData | null
-	loadError: Error | null
-	saveConfig: SaveApiConfigData
-} {
-	const saveConfig = useCallback(async (config: Partial<ApiConfigData>): Promise<void> => {
-		const newConfig = await electronApi.saveConfig(config)
-
-		setCurrentConfig(newConfig)
-	}, [])
-
-	const [currentConfig, setCurrentConfig] = useState<ApiConfigData | null>(null)
-
-	const getConfigFn = useCallback(async () => electronApi.getConfig(), [])
-	const loadData = usePoller<ApiConfigData>(CONFIG_POLL_INTERVAL, getConfigFn)
-	useEffect(() => {
-		if (loadData.data) {
-			setCurrentConfig(loadData.data)
-		}
-	}, [loadData.data])
-
-	return {
-		currentConfig,
-		loadError: loadData.error ?? null,
-		saveConfig,
-	}
-}
 
 export function ElectronApp() {
 	const restConfigApi = useIpcConfigApi()
