+++ webui/src/WebApp.tsx
@@ -7,15 +7,23 @@ import { ConnectionStatus } from './ConnectionStatus'
 import { MyErrorBoundary } from './Util/ErrorBoundary'
 import { useRestConfigApi } from './Api/rest'
 import { useCallback } from 'react'
-import { useFetchInterval } from './Util/useFetchInterval'
 import type { ApiStatusResponse } from '../../src/apiTypes'
+import { usePoller } from './Util/usePoller'
 
 const STATUS_POLL_INTERVAL = 2000
 
 export function WebApp() {
 	const restApi = useRestConfigApi()
 
-	const apiStatus = useFetchInterval<ApiStatusResponse>(STATUS_POLL_INTERVAL, '/api/status')
+	const fetchStatus = useCallback(async () => {
+		const response = await fetch('/api/status')
+		if (!response.ok) {
+			throw new Error(response.statusText)
+		}
+
+		return (await response.json()) as ApiStatusResponse
+	}, [])
+	const apiStatus = usePoller<ApiStatusResponse>(STATUS_POLL_INTERVAL, fetchStatus)
 
 	const rescanSurfaces = useCallback(() => {
 		fetch('/api/rescan', {
