--- webui/src/renderer.d.ts
@@ -1,1 +1,0 @@
-declare const electronApi: typeof import('../../dist/electronPreload').electronApi
