+++ webui/src/vite-env.d.ts
@@ -1,1 +1,3 @@
 /// <reference types="vite/client" />
+
+declare const electronApi: typeof import('../../dist/electronPreload').electronApi
